requirements.txt
torch>=2.0
transformers>=4.40
huggingface_hub>=0.23
datasets>=2.19
numpy>=1.24
pandas>=2.0
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
wandb>=0.16
